create xorg role
Refactor ansible code to use roles.

For now, the idea is that a system (e.g. desktop running Arch Linux) is
configured using a playbook with the roles (e.g. xorg) it needs. This
way systems are configured by a play, using variables and tasks it
needs to set up the system.